Resolve "Send spent in hours"

Closes #35

See merge request junte/spec-projector/backend!21

+++ server/apps/projects/graphql/types/issue.py
@@ -11,7 +11,7 @@ class IssueType(graphene.ObjectType):
     title = graphene.String()
     state = graphene.String()
     due_date = graphene.String()
-    spent = graphene.Int()
+    spent = graphene.Float()
     assignee = graphene.Field(AssigneeType)
 
     class Meta:
